<Refactor> 메인페이지 코드 구조 변경
